requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
pillow>=9.3.0
matplotlib>=3.7.0
tqdm>=4.65.0
PyYAML>=6.0
tensorboard>=2.13.0
pytest>=7.3.1
black>=23.3.0
isort>=5.12.0
flake8>=6.0.0
scipy>=1.10.0
einops>=0.6.1
wandb>=0.15.0  # for experiment tracking
torchmetrics>=1.0.0 